feat: create seperate folder to hold search feature
