+++ src/components/Header.jsx
@@ -1,18 +1,13 @@
 import { useState } from "react"
+import Search from "../features/search/Search"
 
 const Header = () => {
-  const [searchInput, setSearchInput] = useState("")
 
-  const handleChange = ({target}) => {
-    setSearchInput(target.value)
-  }
 
   return (
     <header>
       <h2>Reddit Client App</h2>
-      <label htmlFor="search">
-        <input name='search' id="search" placeholder="Search" onChange={handleChange} value={searchInput} />
-      </label>
+      <Search />
     </header>
   )
 }
